Modal moved to common folder
